+++ src/lib/strict-mode.js
@@ -7,10 +7,10 @@ function createMessage(message) {
 }
 
 module.exports.ERROR = {
-    NO_RETURN_IN_JSDOC: createMessage('Return expression isn\'t covered by jsDoc comment, please, provide more info.'),
-    NO_ARGUMENT_IN_JSDOC: createMessage('This argument isn\'t covered by jsDoc comment, please, provide more info.'),
-    UNUSED_ARGUMENTS_IN_JSDOC: createMessage('Parameters described in JSDoc doesn\'t appear in function signature.'),
-    EMPTY_RETURN_IN_FUNCTION: createMessage('Return statement is empty, but must provide data with specified type.')
+    NO_RETURN_IN_JSDOC: createMessage('Return statement type annotation missing.'),
+    NO_ARGUMENT_IN_JSDOC: createMessage('Function argument type annotation missing.'),
+    UNUSED_ARGUMENTS_IN_JSDOC: createMessage('Arguments described in type annotation doesn\'t appear in function signature.'),
+    EMPTY_RETURN_IN_FUNCTION: createMessage('Return statement must return something.')
 };
 
 /**