+++ test/data/directive-parsing/src/defaults-global-directive-before-function.js
@@ -1,0 +1,5 @@
+// @typecheck
+
+function test(a) {
+    return a;
+}